Split gmail specifics and general google api into separate classes.
